requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2019.9.11        # via requests, sentry-sdk
chardet==3.0.4            # via requests
django-anymail==7.1.0     # via django-ilmoitin
django-appconf==1.0.4     # via django-axes
django-axes==5.3.1        # via -r requirements.in
django-cors-headers==3.1.1  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-helusers==0.5.6    # via -r requirements.in
django-ilmoitin==0.4.0    # via -r requirements.in
django-ipware==2.1.0      # via django-axes
django-mailer==2.0.1      # via django-ilmoitin
django-parler==1.9.2      # via django-ilmoitin
django==2.2.13            # via -r requirements.in, django-anymail, django-appconf, django-axes, django-cors-headers, django-helusers, django-ilmoitin, django-mailer, drf-oidc-auth
djangorestframework==3.10.3  # via -r requirements.in, drf-oidc-auth
drf-oidc-auth==0.9        # via django-helusers
ecdsa==0.13.3             # via python-jose
factory-boy==2.12.0       # via -r requirements.in
faker==2.0.4              # via factory-boy
future==0.17.1            # via pyjwkest, python-jose
idna==2.8                 # via requests
jinja2==2.11.2            # via django-ilmoitin
lockfile==0.12.2          # via django-mailer
markupsafe==1.1.1         # via jinja2
psycopg2==2.8.4           # via -r requirements.in
pyasn1==0.4.7             # via rsa
pycryptodomex==3.9.0      # via pyjwkest
pyjwkest==1.4.2           # via drf-oidc-auth
python-dateutil==2.8.1    # via faker
python-jose==3.0.1        # via django-helusers
pytz==2019.3              # via django
requests==2.22.0          # via django-anymail, django-helusers, pyjwkest
rsa==4.0                  # via python-jose
sentry-sdk==0.12.3        # via -r requirements.in
six==1.12.0               # via django-anymail, django-mailer, faker, pyjwkest, python-dateutil, python-jose
sqlparse==0.3.0           # via django
text-unidecode==1.3       # via faker
urllib3==1.25.6           # via requests, sentry-sdk
